requirements.txt
yfinance
pandas-ta
pandas
numpy
feedparser
requests
openai
python-dotenv
